grafana/requirements.txt
azure-storage-blob==12.19.0
influxdb-client==1.38.0
python-dotenv==1.0.0

